Remove Order template from elements (geometry)
