+++ src/Caribou/Geometry/Base/BaseSegment.h
