+++ src/Caribou/Geometry/Base/BaseTriangle.h
