+++ unittest/Caribou/Geometry/test_hexahedron20.h
@@ -3,16 +3,15 @@
 #include <Eigen/Dense>
 #include <Caribou/macros.h>
 #include <Caribou/constants.h>
-#include <Caribou/Geometry/Hexahedron.h>
-#include <Caribou/Geometry/Tetrahedron.h>
+#include <Caribou/Geometry/Hexahedron20.h>
+#include <Caribou/Geometry/Tetrahedron10.h>
 
-TEST(Hexahedron, Linear) {
+TEST(Hexahedron, Quadratic) {
     using namespace caribou;
-    using Hexahedron = caribou::geometry::Hexahedron;
-    using Tetrahedron = caribou::geometry::Tetrahedron;
-    using Quad = caribou::geometry::Quad<3>;
-    using Triangle = caribou::geometry::Triangle<3>;
-    using Edge = caribou::geometry::Segment<3>;
+    using Hexahedron = caribou::geometry::Hexahedron20;
+    using Tetrahedron = caribou::geometry::Tetrahedron10;
+    using Quad = caribou::geometry::Quad8<3>;
+    using Edge = caribou::geometry::Segment3<3>;
     using LocalCoordinates = Hexahedron::LocalCoordinates;
     using WorldCoordinates = Hexahedron::WorldCoordinates;
 
@@ -37,20 +36,15 @@ TEST(Hexahedron, Linear) {
     }
 
     {
-        Hexahedron h (
-            WorldCoordinates(  -1,  1,   -1), WorldCoordinates(   1,  1,   -1),
-            WorldCoordinates(   1,  1,    1), WorldCoordinates(  -1,  1,    1),
-            WorldCoordinates(-0.5, -1, -0.5), WorldCoordinates( 0.5, -1, -0.5),
-            WorldCoordinates( 0.5, -1,  0.5), WorldCoordinates(-0.5, -1,  0.5)
-        );
+        Hexahedron h;
 
         // Interpolation
-        Eigen::Matrix<FLOATING_POINT_TYPE, 8, 1> values;
-        values << p1(h.node(0)), p1(h.node(1)), p1(h.node(2)), p1(h.node(3)),
-                  p1(h.node(4)), p1(h.node(5)), p1(h.node(6)), p1(h.node(7));
+        Eigen::Matrix<FLOATING_POINT_TYPE, 20, 1> values;
+        values << p2(h.node( 0)), p2(h.node( 1)), p2(h.node( 2)), p2(h.node( 3)), p2(h.node( 4)), p2(h.node( 5)), p2(h.node( 6)), p2(h.node( 7)), p2(h.node( 8)), p2(h.node( 9)),
+                  p2(h.node(10)), p2(h.node(11)), p2(h.node(12)), p2(h.node(13)), p2(h.node(14)), p2(h.node(15)), p2(h.node(16)), p2(h.node(17)), p2(h.node(18)), p2(h.node(19));
         for (const auto & gauss_node : h.gauss_nodes()) {
             const auto x = gauss_node.position;
-            EXPECT_NEAR(h.interpolate(x, values), p1(h.world_coordinates(x)), 1e-10);
+            EXPECT_DOUBLE_EQ(h.interpolate(x, values), p2(h.world_coordinates(x)));
         }
 
         // Inverse transformation
@@ -69,7 +63,7 @@ TEST(Hexahedron, Linear) {
             for (UNSIGNED_INTEGER_TYPE face_id = 0; face_id < h.number_of_boundary_elements(); ++face_id) {
                 Quad face = h.boundary_element(face_id);
                 for (UNSIGNED_INTEGER_TYPE gauss_node_id = 0; gauss_node_id < face.number_of_gauss_nodes();++gauss_node_id) {
-                    inside_points.push_back(h.local_coordinates(face.world_coordinates(face.gauss_node(gauss_node_id).position)));
+                    inside_points.push_back(h.gauss_node(gauss_node_id).position);
                 }
                 for (UNSIGNED_INTEGER_TYPE edge_id = 0; edge_id < face.number_of_boundary_elements(); ++edge_id) {
                     Edge edge = face.boundary_element(edge_id);
@@ -94,7 +88,9 @@ TEST(Hexahedron, Linear) {
             std::vector<LocalCoordinates> outside_points = {};
             for (UNSIGNED_INTEGER_TYPE node_id = 0; node_id < h.number_of_nodes();++node_id) {
                 const WorldCoordinates center_to_point = h.node(node_id) - h.center();
-                outside_points.emplace_back(h.center() + center_to_point*1.01);
+                const auto length = center_to_point.norm();
+                const auto n = center_to_point / length;
+                outside_points.emplace_back(h.center() + n*length*1.01);
             }
             for (UNSIGNED_INTEGER_TYPE face_id = 0; face_id < h.number_of_boundary_elements(); ++face_id) {
                 Quad face = h.boundary_element(face_id);
@@ -125,7 +121,7 @@ TEST(Hexahedron, Linear) {
             const auto x = gauss_node.position;
             const auto w = gauss_node.weight;
             const auto detJ = h.jacobian(x).determinant();
-            numerical_solution_hexa += p1(h.world_coordinates(x)) * w * detJ;
+            numerical_solution_hexa += p2(h.world_coordinates(x)) * w * detJ;
         }
         // Six tetrahedrons
         FLOATING_POINT_TYPE numerical_solution_tetras = 0;
@@ -140,24 +136,10 @@ TEST(Hexahedron, Linear) {
                 const auto x = gauss_node.position;
                 const auto w = gauss_node.weight;
                 const auto detJ = t.jacobian(x).determinant();
-                numerical_solution_tetras += p1(t.world_coordinates(x)) * w * detJ;
+                numerical_solution_tetras += p2(t.world_coordinates(x)) * w * detJ;
             }
         }
 
         EXPECT_DOUBLE_EQ(numerical_solution_hexa, numerical_solution_tetras);
-
-        // Intersection with triangles (test taken from a failing scene)
-        {
-            Hexahedron h (
-                    WorldCoordinates(194.948, -111.776, -1547.85), WorldCoordinates(195.837, -111.776, -1547.85),
-                    WorldCoordinates(195.837, -110.461, -1547.85), WorldCoordinates(194.948, -110.461, -1547.85),
-                    WorldCoordinates(194.948, -111.776, -1546.86), WorldCoordinates(195.837, -111.776, -1546.86),
-                    WorldCoordinates(195.837, -110.461, -1546.86), WorldCoordinates(194.948, -110.461, -1546.86)
-            );
-
-            Triangle t (WorldCoordinates {195.402, -112.092, -1543.81}, WorldCoordinates {196.468, -108.742, -1546.42}, WorldCoordinates {193.869, -110.918, -1548.45});
-
-            ASSERT_TRUE(h.intersects(t, 0));
-        }
     }
 }
