+++ unittest/Caribou/Geometry/test_quad8.h
@@ -6,9 +6,9 @@
 #include <Caribou/macros.h>
 #include <Caribou/constants.h>
 #include <Caribou/Geometry/Triangle.h>
-#include <Caribou/Geometry/Quad.h>
+#include <Caribou/Geometry/Quad8.h>
 
-TEST(Quad, Linear) {
+TEST(Quad, Quadratic) {
     using namespace caribou;
     using std::cos;
     using std::sin;
@@ -17,8 +17,8 @@ TEST(Quad, Linear) {
     // 2D
     {
         using Rotation = Eigen::Matrix<FLOATING_POINT_TYPE, 2, 2>;
-        using Triangle = caribou::geometry::Triangle<_2D>;
-        using Quad = caribou::geometry::Quad<_2D>;
+        using Triangle = caribou::geometry::Triangle6<_2D>;
+        using Quad = caribou::geometry::Quad8<_2D>;
 
         using LocalCoordinates = Quad::LocalCoordinates;
         using WorldCoordinates = Quad::WorldCoordinates;
@@ -34,16 +34,24 @@ TEST(Quad, Linear) {
         EXPECT_DOUBLE_EQ(q.center()[0], 0);
         EXPECT_DOUBLE_EQ(q.center()[1], 0);
 
+        // Inverse transformation
+        for (const auto & gauss_node : q.gauss_nodes()) {
+            EXPECT_MATRIX_NEAR(gauss_node.position, q.local_coordinates(q.world_coordinates(gauss_node.position)), 1e-5);
+        }
+        for (UNSIGNED_INTEGER_TYPE node_id = 0; node_id < q.number_of_nodes();++node_id) {
+            EXPECT_MATRIX_NEAR(q.node(node_id), q.world_coordinates(q.local_coordinates(q.node(node_id))), 1e-5);
+        }
+
         // Contains point
         {
             FLOATING_POINT_TYPE epsilon = 0.000001;
             // Test that all the following points are INSIDE the element
             std::vector<LocalCoordinates> inside_points = {
                 q.local_coordinates(q.center()),
-                q.local_coordinates(q.world_coordinates({-1-epsilon/2., 0})),
-                q.local_coordinates(q.world_coordinates({+1+epsilon/2., 0})),
-                q.local_coordinates(q.world_coordinates({0, +1+epsilon/2})),
-                q.local_coordinates(q.world_coordinates({0, -1-epsilon/2}))
+                q.local_coordinates(q.world_coordinates({-1., 0.5})),
+                q.local_coordinates(q.world_coordinates({+1, -0.5})),
+                q.local_coordinates(q.world_coordinates({0.5, +1})),
+                q.local_coordinates(q.world_coordinates({-0.5, -1}))
             };
             for (UNSIGNED_INTEGER_TYPE node_id = 0; node_id < q.number_of_nodes();++node_id) {
                 inside_points.push_back(q.local_coordinates(q.node(node_id)));
@@ -69,15 +77,12 @@ TEST(Quad, Linear) {
         }
 
         // Interpolation
-        Eigen::Matrix<FLOATING_POINT_TYPE, 4, 1> values (p1(q.node(0)), p1(q.node(1)), p1(q.node(2)), p1(q.node(3)));
+        Eigen::Matrix<FLOATING_POINT_TYPE, 8, 1> values;
+        values << p2(q.node(0)), p2(q.node(1)), p2(q.node(2)), p2(q.node(3)),
+                  p2(q.node(4)), p2(q.node(5)), p2(q.node(6)), p2(q.node(7));
         for (const auto & gauss_node : q.gauss_nodes()) {
             const auto x = gauss_node.position;
-            EXPECT_DOUBLE_EQ(q.interpolate(x, values), p1(q.world_coordinates(x)));
-        }
-
-        // Inverse transformation
-        for (const auto & gauss_node : q.gauss_nodes()) {
-            EXPECT_MATRIX_NEAR(gauss_node.position, q.local_coordinates(q.world_coordinates(gauss_node.position)), 1e-5);
+            EXPECT_NEAR(q.interpolate(x, values), p2(q.world_coordinates(x)), 1e-10);
         }
 
         // Integration
@@ -87,7 +92,7 @@ TEST(Quad, Linear) {
             const auto x = gauss_node.position;
             const auto w = gauss_node.weight;
             const auto detJ = q.jacobian(x).determinant();
-            numerical_solution_quad += p1(q.world_coordinates(x)) * w * detJ;
+            numerical_solution_quad += p2(q.world_coordinates(x)) * w * detJ;
         }
 
         // Two triangles
@@ -99,15 +104,14 @@ TEST(Quad, Linear) {
                 const auto x = gauss_node.position;
                 const auto w = gauss_node.weight;
                 const auto detJ = t.jacobian(x).determinant();
-                numerical_solution_triangles += p1(t.world_coordinates(x)) * w * detJ;
+                numerical_solution_triangles += p2(t.world_coordinates(x)) * w * detJ;
             }
         }
 
         EXPECT_DOUBLE_EQ(numerical_solution_quad, numerical_solution_triangles);
 
         // Frame extraction
-        Eigen::Matrix<FLOATING_POINT_TYPE, 4, 2> nodes;
-        nodes << -25, -25, 25, -25, 25, 25, -25,  25;
+        Quad initial_quad (WorldCoordinates(-25, -25), WorldCoordinates( 25, -25), WorldCoordinates(25, 25), WorldCoordinates(-25, 25));
 
         bool test_failed = false;
         std::array<FLOATING_POINT_TYPE, 8> angles {pi/6, pi/4, pi/3, pi/2, 3*pi/4, pi, 5*pi/4, 7*pi/4};
@@ -116,11 +120,11 @@ TEST(Quad, Linear) {
         for (const auto & a : angles) {
             Rotation R;
             R << cos(a), -sin(a),
-                 sin(a),  cos(a);
+                sin(a),  cos(a);
 
-            Eigen::Matrix<FLOATING_POINT_TYPE, 4, 2> rotated_nodes;
-            for (unsigned int i = 0; i < 4; ++i)
-                rotated_nodes.row(i) = R*nodes.row(i).transpose();
+            Eigen::Matrix<FLOATING_POINT_TYPE, 8, 2> rotated_nodes;
+            for (unsigned int i = 0; i < 8; ++i)
+                rotated_nodes.row(i) = R*initial_quad.node(i);
 
             Quad rotated_quad(rotated_nodes);
             Rotation F = rotated_quad.frame({0, 0});
@@ -138,9 +142,8 @@ TEST(Quad, Linear) {
 
     // 3D
     {
-        using Rotation = Eigen::Matrix<FLOATING_POINT_TYPE, 3, 3>;
-        using Triangle = caribou::geometry::Triangle<_3D>;
-        using Quad = caribou::geometry::Quad<_3D>;
+        using Triangle = caribou::geometry::Triangle6<_3D>;
+        using Quad = caribou::geometry::Quad8<_3D>;
 
         using LocalCoordinates = Quad::LocalCoordinates;
         using WorldCoordinates = Quad::WorldCoordinates;
@@ -157,16 +160,24 @@ TEST(Quad, Linear) {
         EXPECT_DOUBLE_EQ(q.center()[1], 0);
         EXPECT_DOUBLE_EQ(q.center()[2], 0);
 
+        // Inverse transformation
+        for (const auto & gauss_node : q.gauss_nodes()) {
+            EXPECT_MATRIX_NEAR(gauss_node.position, q.local_coordinates(q.world_coordinates(gauss_node.position)), 1e-5);
+        }
+        for (UNSIGNED_INTEGER_TYPE node_id = 0; node_id < q.number_of_nodes();++node_id) {
+            EXPECT_MATRIX_NEAR(q.node(node_id), q.world_coordinates(q.local_coordinates(q.node(node_id))), 1e-5);
+        }
+
         // Contains point
         {
             FLOATING_POINT_TYPE epsilon = 0.000001;
             // Test that all the following points are INSIDE the element
             std::vector<LocalCoordinates> inside_points = {
                 q.local_coordinates(q.center()),
-                q.local_coordinates(q.world_coordinates({-1-epsilon/2., 0})),
-                q.local_coordinates(q.world_coordinates({+1+epsilon/2., 0})),
-                q.local_coordinates(q.world_coordinates({0, +1+epsilon/2})),
-                q.local_coordinates(q.world_coordinates({0, -1-epsilon/2}))
+                q.local_coordinates(q.world_coordinates({-1., 0.5})),
+                q.local_coordinates(q.world_coordinates({+1, -0.5})),
+                q.local_coordinates(q.world_coordinates({0.5, +1})),
+                q.local_coordinates(q.world_coordinates({-0.5, -1}))
             };
             for (UNSIGNED_INTEGER_TYPE node_id = 0; node_id < q.number_of_nodes();++node_id) {
                 inside_points.push_back(q.local_coordinates(q.node(node_id)));
@@ -192,15 +203,12 @@ TEST(Quad, Linear) {
         }
 
         // Interpolation
-        Eigen::Matrix<FLOATING_POINT_TYPE, 4, 1> values (p1(q.node(0)), p1(q.node(1)), p1(q.node(2)), p1(q.node(3)));
+        Eigen::Matrix<FLOATING_POINT_TYPE, 8, 1> values;
+        values << p2(q.node(0)), p2(q.node(1)), p2(q.node(2)), p2(q.node(3)),
+                  p2(q.node(4)), p2(q.node(5)), p2(q.node(6)), p2(q.node(7));
         for (const auto & gauss_node : q.gauss_nodes()) {
             const auto x = gauss_node.position;
-            EXPECT_DOUBLE_EQ(q.interpolate(x, values), p1(q.world_coordinates(x)));
-        }
-
-        // Inverse transformation
-        for (const auto & gauss_node : q.gauss_nodes()) {
-            EXPECT_MATRIX_NEAR(gauss_node.position, q.local_coordinates(q.world_coordinates(gauss_node.position)), 1e-5);
+            EXPECT_DOUBLE_EQ(q.interpolate(x, values), p2(q.world_coordinates(x)));
         }
 
         // Integration
@@ -211,7 +219,7 @@ TEST(Quad, Linear) {
             const auto w = gauss_node.weight;
             const auto J = q.jacobian(x);
             const auto detJ = J.col(0).cross(J.col(1)).norm();
-            numerical_solution_quad += p1(q.world_coordinates(x)) * w * detJ;
+            numerical_solution_quad += p2(q.world_coordinates(x)) * w * detJ;
         }
 
         // Two triangles
@@ -224,48 +232,10 @@ TEST(Quad, Linear) {
                 const auto w = gauss_node.weight;
                 const auto J = t.jacobian(x);
                 const auto detJ = J.col(0).cross(J.col(1)).norm();
-                numerical_solution_triangles += p1(t.world_coordinates(x)) * w * detJ;
+                numerical_solution_triangles += p2(t.world_coordinates(x)) * w * detJ;
             }
         }
 
         EXPECT_DOUBLE_EQ(numerical_solution_quad, numerical_solution_triangles);
-
-        // Frame extraction
-        Eigen::Matrix<FLOATING_POINT_TYPE, 4, 3> nodes;
-        nodes << -25, -25, 0, 25, -25, 0, 25, 25, 0, -25,  25, 0;
-
-        bool test_failed = false;
-        std::array<FLOATING_POINT_TYPE, 8> angles {pi/6, pi/4, pi/3, pi/2, 3*pi/4, pi, 5*pi/4, 7*pi/4};
-        Eigen::Matrix<FLOATING_POINT_TYPE, 3, 3*angles.size()> results; // Should be 8 3x3 Identity matrices
-        unsigned int angle_id = 0;
-        for (const auto & a : angles) {
-            Rotation R, Rx, Ry, Rz;
-            Rx << 1,       0,       0,
-                  0,       cos(a), -sin(a),
-                  0,       sin(a),  cos(a);
-            Ry << cos(a),  0,       sin(a),
-                  0,       1,       0,
-                  -sin(a), 0,       cos(a);
-            Rz << cos(a),  -sin(a), 0,
-                  sin(a),  cos(a),  0,
-                  0      , 0,       1;
-            R = Rz*Ry*Rx;
-
-            Eigen::Matrix<FLOATING_POINT_TYPE, 4, 3> rotated_nodes;
-            for (unsigned int i = 0; i < 4; ++i)
-                rotated_nodes.row(i) = R*nodes.row(i).transpose();
-
-            Quad rotated_quad(rotated_nodes);
-            Rotation F = rotated_quad.frame({0, 0});
-            results.block<3,3>(0, 3*angle_id) = F.transpose()*R;
-            Rotation diff = R - F;
-            if (not (IN_CLOSED_INTERVAL(-1e-15, diff.minCoeff(), 1e-15) and
-                     IN_CLOSED_INTERVAL(-1e-15, diff.maxCoeff(), 1e-15))) {
-                test_failed = true;
-            }
-            ++angle_id;
-        }
-        Eigen::IOFormat clean(2, 0, ", ", "\n", "[", "]");
-        EXPECT_FALSE(test_failed) << results.format(clean);
     }
 }
